requirements.txt
snowflake-connector-python>=3.6.0
python-dotenv>=1.0.1
composio-core>=0.7.0
composio-crewai>=0.7.0
crewai>=0.28.0
